Refactor Plaid scripts into global server
