--- src/scripts/testSheets.js
@@ -1,10 +1,0 @@
-require('dotenv').config();
-
-const { updateSheet } = require('../lib/google/sheets');
-
-updateSheet([
-  {
-    range: 'A1',
-    values: [['It worked!']]
-  }
-]);
