+++ src/views/plaid.ejs
@@ -5,7 +5,7 @@
     <title>Plaid Link</title>
     <link rel="stylesheet" href="https://threads.plaid.com/threads.css" />
 
-    <link rel="stylesheet" type="text/css" href="plaid.css" />
+    <link rel="stylesheet" type="text/css" href="/static/plaid.css" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
   </head>
   <body>
